backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yidao-backend"
version = "0.1.0"
description = "YiDao 易道 — 对话式生活分析与教练助手 API 后端"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "gunicorn>=23.0.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "asyncpg>=0.31.0",
    "psycopg[binary]>=3.3.0",  # Sync driver for Celery
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "redis>=4.2.0,<5.0.0",  # Pinned for fastapi-cache2 compatibility
    "celery>=5.6.0",
    "email-validator>=2.1.0",
    "python-multipart>=0.0.6",
    "alembic>=1.18.0",
    # Monitoring & Observability
    "prometheus-client>=0.24.0",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-sdk>=1.39.0",
    "opentelemetry-exporter-otlp>=1.39.0",
    "opentelemetry-instrumentation-fastapi>=0.60b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b0",
    "opentelemetry-instrumentation-redis>=0.60b0",
    "opentelemetry-instrumentation-logging>=0.60b0",
    "opentelemetry-instrumentation-celery>=0.60b0",
    # Rate Limiting
    "slowapi>=0.1.9",
    # Caching
    "fastapi-cache2[redis]>=0.2.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    # AI / LLM (latest stable: langgraph 1.x, langchain 1.x, langfuse 3.x)
    "langgraph>=1.0.0,<2.0.0",
    "langchain-core>=1.2.0,<2.0.0",
    "langchain-openai>=1.1.0,<2.0.0",
    "langgraph-checkpoint-postgres>=3.0.0,<4.0.0",
    "langfuse>=3.0.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "httpx>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, with dependencies)",
    "api: API integration tests",
    "e2e: End-to-end tests (slowest, full stack)",
    "performance: Performance tests",
    "security: Security tests",
    "slow: Slow running tests",
    "database: Tests requiring database",
    "redis: Tests requiring Redis",
]
